+++ ATinyQC3Controller/mod/imath/imath.h
@@ -5,6 +5,8 @@
  *  Author: IR
  */
 
+#include <avr/io.h>
+
 #ifndef IMATH_H_
 #define IMATH_H_
 
@@ -21,6 +23,6 @@ _a < _b ? _a : _b; })
 #define clamp(val, minV, maxV) max(min(val, maxV), minV)
 
 // Exponentiation by squaring
-int ipow(int base, int exp);
+uint16_t ipow(uint16_t base, uint16_t exp);
 
 #endif /* IMATH_H_ */
